docker/requirements.txt
kafka-python==2.0.2
requests==2.28.1
apache-airflow-providers-apache-spark==4.1.0